requirements.txt
langchain==0.1.0
langchain-community
#langchain-groq
langchain-ollama
langchain-openai==0.0.2
python-dotenv==1.0.0
openai==1.3.0
slack-sdk==3.21.3